+++ .github/workflows/main.yml
@@ -12,6 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
+
       - name: auto deploy steps
         uses: appleboy/ssh-action@master
         with:
