+++ root/src/index.ejs
@@ -37,6 +37,7 @@
 				"imports": {
 					"single-spa": "https://cdn.jsdelivr.net/npm/single-spa@5.9.0/lib/system/single-spa.min.js",
 
+					"@zuri/plugin-header": "https://staging.zuri.chat/zuri-plugin-header.js",
 					"@zuri/manage-user-modal": "https://staging.zuri.chat/zuri-manageUser.js",
 					"@zuri/zuri-ui": "https://staging.zuri.chat/zuri-zuri-ui.js",
 					"@zuri/utilities": "https://staging.zuri.chat/zuri-utilities.js",
