+++ server/client/src/components/musicRoom.tsx
@@ -1,6 +1,7 @@
 import styled from "styled-components";
 import { ToastContainer } from "react-toastify";
-import { headerConfig, PluginHeader } from "../utils/config";
+import Parcel from "single-spa-react/parcel";
+import { pluginHeader, headerConfig } from "../utils/config";
 import Playlist from "./playlist";
 import PasteUrl from "./common/pasteUrl";
 import EnterRoomModal from "./modals/enterRoom";
@@ -9,11 +10,7 @@ import { selectShowPasteUrl } from "../app/uiSlice";
 import { useEffect, useState } from "react";
 import userService from "../services/userService";
 // import MessageBoard from "./messageBoard";
-import {
-	selectIsMember,
-	setCurrentUser,
-	setMembership
-} from "../app/usersSlice";
+import { selectIsMember, setMembership } from "../app/usersSlice";
 import eventService from "../services/eventService";
 import Loader from "./loader";
 
@@ -35,7 +32,7 @@ function MusicRoom() {
 	}, [reload, isMember]);
 
 	useEffect(() => {
-		userService.getCurrentUser().then(user => dispatch(setCurrentUser(user)));
+		// userService.getCurrentUser().then(user => dispatch(setCurrentUser(user)));
 
 		eventService.connect();
 	}, []);
@@ -59,7 +56,12 @@ function MusicRoom() {
 					/>
 				</div>
 
-				<PluginHeader {...headerConfig(members, () => setReload(!reload))} />
+				<Parcel
+					config={pluginHeader}
+					// wrapWith="div"
+					// wrapStyle={{ width: "100%" }}
+					headerConfig={headerConfig(members, () => setReload(!reload))}
+				/>
 
 				<div>
 					<Playlist />
