+++ server/client/src/services/httpService.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 
-const org_id = localStorage.getItem("currentWorkspace") || "61695d8bb2cc8a9af4833d46";
-const room_id = sessionStorage.getItem("currentRoom") || "6169d8b54bfde011fe582e65";
+const org_id = "61695d8bb2cc8a9af4833d46";
+const room_id = "6169d8b54bfde011fe582e65";
 
 // axios.defaults.baseURL = "http://localhost:8000/music/api/v1";
 // axios.defaults.baseURL = "https://music.zuri.chat/music/api/v1/";
@@ -13,7 +13,7 @@ export const endpoints = {
 	currentsong: "songs/current",
 	deletesong: "songs/delete",
 	songsearch: "songsearch",
-	likesong: "songs/like",
+	likesong: "songs/likecount",
 
 	// Comments endpoints
 	comments: "comments",
