+++ server/client/src/utils/config.ts
@@ -1,9 +1,9 @@
 // @ts-ignore
-import { PluginHeader } from "@zuri/zuri-ui";
+import { pluginHeader } from "@zuri/plugin-header";
 import icon from "../media/musicRoomIcon.svg";
 import userService from "../services/userService";
 
-export { PluginHeader };
+export { pluginHeader };
 
 export const headerConfig = (members: User[], cb) => {
 	return {
