+++ server/config/urls.py
@@ -5,7 +5,7 @@ from django.urls import include, path
 from django.views.generic import TemplateView
 from drf_spectacular.views import (SpectacularAPIView, SpectacularRedocView,
                                    SpectacularSwaggerView)
-from music.view.static import SidebarView
+from music.views import SidebarView
 
 urlpatterns = [
     # coming from frontend app using react every thing we connect in App.js in components will be
