+++ server/music/models.py
@@ -5,10 +5,6 @@ from dataclasses import dataclass, field
 class Song:  # for the songs
 
     _id: str
-    plugin_id: str
-    organization_id: str
-    collection_name: str
-    room_id: str
     title: str
     duration: str
     albumCover: str
@@ -30,10 +26,6 @@ class songLikeCount:
 class Room:  # for the rooms
 
     _id: str
-    plugin_id: str
-    organization_id: str
-    collection_name: str
-    plugin_name: str
     room_name: str
     description: str
     created_by: str
@@ -65,10 +57,6 @@ class Emojis:  # for the emojis
 class Chat:  # for the chats
 
     _id: str
-    plugin_id: str
-    organization_id: str
-    collection_name: str
-    room_id: str
     username: str
     userId: str
     time: int
